refactor: cleanup
