+++ src/lib/characterPreview/ShowcaseLightCone.tsx
@@ -1,7 +1,6 @@
 import { Flex, Typography } from 'antd'
-import { ShowcaseDisplayDimensions } from 'lib/characterPreview/CharacterPreview'
 import { showcaseDropShadowFilter, showcaseOutline, showcaseShadow, ShowcaseSource } from 'lib/characterPreview/CharacterPreviewComponents'
-import { ShowcaseMetadata } from 'lib/characterPreview/characterPreviewController'
+import { ShowcaseDisplayDimensions, ShowcaseMetadata } from 'lib/characterPreview/characterPreviewController'
 import StatText from 'lib/characterPreview/StatText'
 import { middleColumnWidth, parentW } from 'lib/constants/constantsUi'
 import { LoadingBlurredImage } from 'lib/ui/LoadingBlurredImage'
