+++ src/lib/characterPreview/ShowcasePortrait.tsx
@@ -1,8 +1,8 @@
 import { EditOutlined } from '@ant-design/icons'
 import { Button, Flex, Typography } from 'antd'
 import CharacterCustomPortrait from 'lib/characterPreview/CharacterCustomPortrait'
-import { ShowcaseDisplayDimensions } from 'lib/characterPreview/CharacterPreview'
 import { showcaseButtonStyle, showcaseDropShadowFilter, showcaseOutline, ShowcaseSource } from 'lib/characterPreview/CharacterPreviewComponents'
+import { ShowcaseDisplayDimensions } from 'lib/characterPreview/characterPreviewController'
 import { parentH, parentW } from 'lib/constants/constantsUi'
 import EditImageModal from 'lib/overlays/modals/EditImageModal'
 import { Assets } from 'lib/rendering/assets'
